poprawki - usunięcie implementacji algorytmu
